Add command line parameters to start applications
This change adds `--kiwix`, `--kolibri`, and `--readme` command line
parameters for Endless Launcher. Using these parameters, it is possible
to open these immediately, for example from a special shortcut, instead
of pressing buttons in the launcher.

https://phabricator.endlessm.com/T31030
